fix parasites tides generator

+++ source/mutableSources32/stmlib/utils/dsp.h
@@ -52,7 +52,7 @@ inline int16_t Interpolate88(const int16_t* table, uint16_t index);
 
 inline int16_t Interpolate1022(const int16_t* table, uint32_t phase);
 
-inline int16_t Interpolate115(const int16_t* table, uint32_t phase);
+inline int16_t Interpolate115(const int16_t* table, uint16_t phase);
 
 inline int16_t Crossfade(
     const int16_t* table_a,
